+++ packages/integration-sdk-runtime/src/execution/__tests__/jobState.test.ts
@@ -1,11 +1,14 @@
 import {
   createStepJobState,
-  DuplicateKeyTracker,
   TypeTracker,
   MemoryDataStore,
   CreateStepJobStateParams,
   TypeTrackerStepSummary,
 } from '../jobState';
+import {
+  DuplicateEntityReport,
+  DuplicateKeyTracker,
+} from '../duplicateKeyTracker';
 import { v4 as uuid } from 'uuid';
 import { FileSystemGraphObjectStore } from '../../storage';
 import { vol } from 'memfs';
@@ -55,12 +58,19 @@ function createInMemoryStepGraphObjectDataUploaderCollector(
 function getMockCreateStepJobStateParams(
   partial?: Partial<CreateStepJobStateParams>,
 ): CreateStepJobStateParams {
+  const onDuplicateEntityKey = (
+    DuplicateEntityReport: DuplicateEntityReport,
+  ) => {
+    return;
+  };
+
   return {
     stepId: uuid(),
     graphObjectStore: new FileSystemGraphObjectStore(),
     duplicateKeyTracker: new DuplicateKeyTracker(),
     typeTracker: new TypeTracker(),
     dataStore: new MemoryDataStore(),
+    onDuplicateEntityKey,
     ...partial,
   };
 }
@@ -480,9 +490,7 @@ describe('#TypeTracker', () => {
     test('should return empty summary when step does not exist', () => {
       const typeTracker = new TypeTracker();
 
-      const expected: TypeTrackerStepSummary = {
-        graphObjectTypeSummary: [],
-      };
+      const expected: TypeTrackerStepSummary = {};
 
       expect(typeTracker.summarizeStep('a')).toEqual(expected);
     });
@@ -521,16 +529,12 @@ describe('#TypeTracker', () => {
       });
 
       const expected: TypeTrackerStepSummary = {
-        graphObjectTypeSummary: [
-          {
-            _type: 'my_type_1',
-            total: 3,
-          },
-          {
-            _type: 'my_type_2',
-            total: 1,
-          },
-        ],
+        my_type_1: {
+          total: 3,
+        },
+        my_type_2: {
+          total: 1,
+        },
       };
 
       expect(typeTracker.summarizeStep('a')).toEqual(expected);