+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to
 
 ## Unreleased
 
+## 8.30.1 - 2023-01-10
+
+### Added
+
+- Added new `DuplicateEntityReport` when an integration errors with a duplicate
+  **\_key**.
+
+## 8.30.0 - 2022-12-20
+
+### Changed
+
+- Step summary refactored to be `summary.$TYPE: { total: number }`. This will
+  enable better logs querying.
+
 ## 8.29.3 - 2022-11-28
 
 ### Fixed